+++ test/e2e/playwright.config.ts
@@ -1,8 +1,8 @@
-import { defineConfig, devices } from '@playwright/test';
+import { defineConfig, devices } from "@playwright/test";
 
 export default defineConfig({
     // Look for test files in the "tests" directory, relative to this configuration file.
-    testDir: '.',
+    testDir: ".",
 
     // Run all tests in parallel.
     fullyParallel: true,
@@ -17,26 +17,26 @@ export default defineConfig({
     workers: 5,
 
     // Reporter to use
-    reporter: 'html',
+    reporter: "html",
 
     use: {
         // Base URL to use in actions like `await page.goto('/')`.
-        baseURL: 'http://127.0.0.1:8182',
+        baseURL: "http://127.0.0.1:8182",
 
         // Collect trace when retrying the failed test.
-        trace: 'on-first-retry',
+        trace: "on-first-retry",
     },
     // Configure projects for major browsers.
     projects: [
         {
-            name: 'chromium',
-            use: { ...devices['Desktop Chrome'] },
+            name: "chromium",
+            use: { ...devices["Desktop Chrome"] },
         },
     ],
     // Run your local dev server before starting the tests.
     webServer: {
-        command: '(cd $(git rev-parse --show-toplevel); make run-e2e 1>&2)',
-        url: 'http://127.0.0.1:8182',
+        command: "(cd $(git rev-parse --show-toplevel); make run-e2e 1>&2)",
+        url: "http://127.0.0.1:8182",
     },
 
     timeout: 15000,