+++ CHANGELOG.md
@@ -6,13 +6,17 @@
 - separate html, css and js by @lirlia in https://github.com/lirlia/prel/pull/39
 
 ## [v0.3.0](https://github.com/lirlia/prel/compare/v0.2.3...v0.3.0) - 2024-01-02
+
 ### Exciting New Features 🎉
+
 - feature: add mobile device css by @lirlia in https://github.com/lirlia/prel/pull/36
+
 ### Other Changes
+
 - add badge in readme by @lirlia in https://github.com/lirlia/prel/pull/32
 - update release.yml by @lirlia in https://github.com/lirlia/prel/pull/31
 - Update Songmu/tagpr action to v1.2.0 by @renovate in https://github.com/lirlia/prel/pull/34
-- fix: uncheck when reject request  by @lirlia in https://github.com/lirlia/prel/pull/35
+- fix: uncheck when reject request by @lirlia in https://github.com/lirlia/prel/pull/35
 
 ## [v0.2.3](https://github.com/lirlia/prel/compare/v0.2.2...v0.2.3) - 2023-12-30
 
@@ -21,6 +25,7 @@
 ## [v0.2.1](https://github.com/lirlia/prel/compare/v0.2.0...v0.2.1) - 2023-12-30
 
 ## [v0.2.0](https://github.com/lirlia/prel/compare/v0.1.1...v0.2.0) - 2023-12-30
+
 - feat: add iam roles filtering by @lirlia in https://github.com/lirlia/prel/pull/5
 - fix license by @lirlia in https://github.com/lirlia/prel/pull/7
 - chage db schema for psqldf by @lirlia in https://github.com/lirlia/prel/pull/9
@@ -34,6 +39,7 @@
 - use e2e db by @lirlia in https://github.com/lirlia/prel/pull/24
 
 ## [v0.1.1](https://github.com/lirlia/prel/compare/v0.1.0...v0.1.1) - 2023-12-27
+
 - Update postgres Docker tag to v16 by @renovate in https://github.com/lirlia/prel/pull/2
 
 ## [v0.1.0](https://github.com/lirlia/prel/commits/v0.1.0) - 2023-12-27